backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
numpy==1.24.3
faiss-cpu==1.7.4
opencv-python-headless==4.8.0.74
insightface==0.7.3
firebase-admin==6.2.0
setuptools>=68.0.0
wheel>=0.40.0
onnxruntime==1.16.0